Division of the VIewMOdel from the VIew according to the principles of MVVM and SoC – Separation of Concern
